# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5ebbbc9ebbf7210dd3bd64a3
  title: meetup
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  /catgory:
    post:
      x-skaffolder-id: 5ebbbd32bbf7210dd3bd6535
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ebbbd12bbf7210dd3bd6529
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Category
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category_name:
                  type: string
    get:
      x-skaffolder-id: 5ebbbd32bbf7210dd3bd653a
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ebbbd12bbf7210dd3bd6529
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Category
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  "/catgory/{id}":
    delete:
      x-skaffolder-id: 5ebbbd32bbf7210dd3bd653b
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebbbd12bbf7210dd3bd6529
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebbbd32bbf7210dd3bd6534
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebbbd12bbf7210dd3bd6529
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    post:
      x-skaffolder-id: 5ebbbd32bbf7210dd3bd6537
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebbbd12bbf7210dd3bd6529
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
...
  /user:
    post:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b6
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b1
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b0
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b7
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b3
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Category:
      x-skaffolder-id: 5ebbbd12bbf7210dd3bd6529
      x-skaffolder-id-db: 5ebbbc9ebbf7210dd3bd64a4
      x-skaffolder-id-entity: 5ebbbd12bbf7210dd3bd6527
      x-skaffolder-url: /catgory
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        category_name:
          type: string
          x-skaffolder-id-attr: 5ebbbd12bbf7210dd3bd6528
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
      x-skaffolder-relations: 
      required:
        - _id
        - category_name
...
    User:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-id-db: 5ebbbc9ebbf7210dd3bd64a4
      x-skaffolder-id-entity: 5ebbbc9ebbf7210dd3bd64a5
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a8
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a9
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a7
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64ab
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64aa
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a6
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...